Automate all the thingz
This makes each PC run in a loop where it requests a read or write
randomly every second.

There is a BIG HUGE bug though because each PC now has it's own vector
clock (as God intended) but there isn't any mechanism to pass the vector
clocks around to the other clients.

-TODO (IMPORTANT): Find a way to pass each PCs vector clock to the
others so that they can update them appropriately.